[skip hazardlib]
